move `NotifyOnDestruction` into `MessageQueueState`
