+++ main/lsp/protocol.cc
@@ -19,18 +19,6 @@ namespace spd = spdlog;
 namespace sorbet::realmain::lsp {
 
 namespace {
-class NotifyOnDestruction {
-    absl::Mutex &mutex;
-    bool &flag;
-
-public:
-    NotifyOnDestruction(absl::Mutex &mutex, bool &flag) : mutex(mutex), flag(flag){};
-    ~NotifyOnDestruction() {
-        absl::MutexLock lck(&mutex);
-        flag = true;
-    }
-};
-
 class TerminateOnDestruction final {
     TaskQueue &queue;
 
@@ -116,7 +104,7 @@ public:
 unique_ptr<Joinable> LSPPreprocessor::runPreprocessor(MessageQueueState &messageQueue, absl::Mutex &messageQueueMutex) {
     return runInAThread("lspPreprocess", [this, &messageQueue, &messageQueueMutex] {
         // Propagate the termination flag across the two queues.
-        NotifyOnDestruction notifyIncoming(messageQueueMutex, messageQueue.terminate);
+        MessageQueueState::NotifyOnDestruction notify(messageQueue, messageQueueMutex);
         TerminateOnDestruction notifyProcessing(*taskQueue);
         owner = this_thread::get_id();
         while (true) {
@@ -194,7 +182,7 @@ optional<unique_ptr<core::GlobalState>> LSPLoop::runLSP(shared_ptr<LSPInput> inp
         runInAThread("lspReader", [&messageQueue, &messageQueueMutex, logger = logger, input = move(input)] {
             // Thread that executes this lambda is called reader thread.
             // This thread _intentionally_ does not capture `this`.
-            NotifyOnDestruction notify(messageQueueMutex, messageQueue.terminate);
+            MessageQueueState::NotifyOnDestruction notify(messageQueue, messageQueueMutex);
             auto timeit = make_unique<Timer>(logger, "getNewRequest");
             while (true) {
                 auto readResult = input->read();
@@ -230,7 +218,7 @@ optional<unique_ptr<core::GlobalState>> LSPLoop::runLSP(shared_ptr<LSPInput> inp
         // Ensure Watchman thread gets unstuck when thread exits prior to initialization.
         NotifyNotificationOnDestruction notify(initializedNotification);
         // Ensure preprocessor, reader, and watchman threads get unstuck when thread exits.
-        NotifyOnDestruction notifyIncoming(messageQueueMutex, messageQueue.terminate);
+        MessageQueueState::NotifyOnDestruction notifyIncoming(messageQueue, messageQueueMutex);
         while (true) {
             unique_ptr<LSPTask> task;
             {
